Relocate NpcNameFinder from core to sharedlib in order to share code with ImageFilter. Remove mouse input based actions from NpcNameFinder
